requirements.txt
torch>=2.0.0
transformers>=4.36.0
huggingface_hub>=0.0.17
llama-cpp-python>=0.2
datasets>=2.14.0
peft>=0.7.0
bitsandbytes>=0.41.0
wandb>=0.15.0
PyYAML>=6.0
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=4.25.0
spacy
flash-attn --no-build-isolation